Update step002_csiinblurout.py
rescale width and height should be integers

+++ python/step002_csiinblurout.py
@@ -32,7 +32,7 @@ while True:
                                 .box_filter(5, border=vpi.Border.ZERO)
         p1_vpiimg1 = vpi_img1.convert(vpi.Format.U8) \
                                 .box_filter(5, border=vpi.Border.ZERO) \
-                                .rescale((vpi_img1.width*0.5, vpi_img1.height*0.5), interp=vpi.Interp.LINEAR, border=vpi.Border.ZERO) \
+                                .rescale((int(vpi_img1.width*0.5), int(vpi_img1.height*0.5)), interp=vpi.Interp.LINEAR, border=vpi.Border.ZERO) \
                                 .box_filter(5, border=vpi.Border.ZERO)
     with p1_vpiimg0.rlock():
         out_img0 = p1_vpiimg0.cpu()
